Monster Race Step2
GameCenter 클래스 삭제
