--- src/com/hsik0225/GameCenter.java
@@ -1,7 +1,0 @@
-package com.hsik0225;
-
-public class GameCenter {
-    public static void main(String[] args) {
-        GameMachine gameMachine = new GameMachine();
-    }
-}
