+++ src/com/hsik0225/Input.java
@@ -3,9 +3,6 @@ package com.hsik0225;
 import java.util.Scanner;
 
 public class Input {
-    private int MONSTER_COUNT;
-    private int CHALLENGE_COUNT;
-
     public int input() {
         Scanner scanner = new Scanner(System.in);
         String inputLine = scanner.nextLine();
@@ -14,13 +11,5 @@ public class Input {
             inputLine = scanner.nextLine();
         }
         return Integer.parseInt(inputLine);
-    }
-
-    public void setMONSTER_COUNT() {
-        this.MONSTER_COUNT = input();
-    }
-
-    public void setCHALLENGE_COUNT() {
-        this.CHALLENGE_COUNT = input();
     }
 }
