+++ src/com/hsik0225/MonsterRace.java
@@ -5,8 +5,22 @@ import java.util.Scanner;
 import java.util.stream.IntStream;
 
 public class MonsterRace {
-    public MonsterRace() {
-        run();
+    private Input input;
+    private Output output;
+
+    public MonsterRace(){
+        input = new Input();
+        output = new Output();
+    }
+
+    public void setGame(){
+        output.startMonsterRace();
+
+        output.askMONSTERS_COUNT();
+        final int MONSTER_COUNT = input.input();
+
+        output.askCHALLENGE_COUNT();
+        final int CHALLENGE_COUNT = input.input();
     }
 
     public void run() {
